+++ test/LibSharp.UnitTests/Caching/LazyAsyncExecutionAndPublicationUnitTests.cs
@@ -142,5 +142,16 @@ public class LazyAsyncExecutionAndPublicationUnitTests
         }
     }
 
+    [TestMethod]
+    public async Task FromFactory_NullTask_ThrowsInvalidOperationException()
+    {
+        // Arrange
+        using LazyAsyncExecutionAndPublication<int> lazy = new LazyAsyncExecutionAndPublication<int>(_ => null);
+
+        // Act
+        _ = await Assert.ThrowsExactlyAsync<InvalidOperationException>(async () =>
+            await lazy.GetValueAsync(CancellationToken.None).ConfigureAwait(false)).ConfigureAwait(false);
+    }
+
     public TestContext TestContext { get; set; }
 }